+++ client/src/components/molecules/FilterSection/FilterSection.types.tsx
@@ -1,0 +1,6 @@
+export default interface FilterSectionPropType {
+  mappingArray: {displayText: string, valueText: string}[];
+  elementName: string;
+  filterValue?: undefined | string;
+  setFilterValue: React.Dispatch<React.SetStateAction<undefined|string>>;
+}
