+++ client/src/components/pages/Home/data/data.types.tsx
@@ -1,0 +1,4 @@
+export interface filterArrayType {
+  displayText: string;
+  valueText: string;
+}
